requirements.txt
fastapi==0.115.12
langchain_core==0.3.59
langchain_openai==0.3.16
langgraph==0.4.2
pydantic==2.11.4
python-dotenv==1.1.0
Requests==2.32.3
uvicorn
langchain==0.3.4
python_jose==3.4.0
python-multipart